Removed structs package automatically encoding with protobuf marshalling
Instead the caller of raftApply that wants to opt-in to protobuf encoding will have to call
raftApplyWithEncoder(t, message, proto.EncodeInterface)
