+++ agent/structs/structs.go
@@ -1724,43 +1724,16 @@ func (r *TombstoneRequest) RequestDatacenter() string {
 	return r.Datacenter
 }
 
-type ProtoMarshaller interface {
-	Size() int
-	MarshalTo([]byte) (int, error)
-	Unmarshal([]byte) error
-	ProtoMessage()
-}
-
 // msgpackHandle is a shared handle for encoding/decoding of structs
 var msgpackHandle = &codec.MsgpackHandle{}
 
 // Decode is used to decode a MsgPack encoded object
 func Decode(buf []byte, out interface{}) error {
-	if marshaller, ok := out.(ProtoMarshaller); ok {
-		return marshaller.Unmarshal(buf)
-	}
-	return decodeMsgpack(buf, out)
-}
-
-func decodeMsgpack(buf []byte, out interface{}) error {
 	return codec.NewDecoder(bytes.NewReader(buf), msgpackHandle).Decode(out)
 }
 
 // Encode is used to encode a MsgPack object with type prefix
 func Encode(t MessageType, msg interface{}) ([]byte, error) {
-	if marshaller, ok := msg.(ProtoMarshaller); ok {
-		data := make([]byte, marshaller.Size()+1)
-		data[0] = uint8(t)
-		if _, err := marshaller.MarshalTo(data[1:]); err != nil {
-			return nil, err
-		}
-		return data, nil
-	}
-
-	return encodeMsgpack(t, msg)
-}
-
-func encodeMsgpack(t MessageType, msg interface{}) ([]byte, error) {
 	var buf bytes.Buffer
 	buf.WriteByte(uint8(t))
 	err := codec.NewEncoder(&buf, msgpackHandle).Encode(msg)
